Override filetypes that already exist in filetype.vim
Some OCaml filetypes already exist in NeoVim's filetype.vim/filetype.lua.
`setf` only sets the filetype if it is not already set.
Because filetype.vim already set the filetype for *.mli/*.mll/*.mly and
is loaded before this ftdetect plugin we have to override it using `set ft`.

This is 'solution A' from https://neovim.io/doc/user/filetype.html#new-filetype

The other filetypes are either new, or match what is already declared in
filetype.vim so no need for overrides there.

Signed-off-by: Edwin Török <edwin@etorok.net>